Feat: Create Navbar - 22.05.23
