--- src/App.css
@@ -1,62 +1,0 @@
-/* MAIN - chris */
-
-.HomePage {
-  display: flex;
-  justify-content: center;
-}
-
-@media (min-width: 1500px) {
-  .MainNav {
-    width: 1400px;
-  }
-}
-
-@media (max-width: 1500px) {
-  .MainNav {
-    width: 90vw;
-  }
-}
-
-.MainNav nav {
-  font-size: 15px;
-  font-weight: bold;
-  align-items: center;
-  padding: 20px;
-  display: flex;
-}
-
-.MainNav a {
-  margin: 20px;
-  align-items: center;
-  color: black;
-  text-decoration: none;
-  justify-content: end;
-}
-
-.MainNav img {
-  display: inline-block;
-  align-items: center;
-  margin-top: 15px;
-  margin-bottom: 15px;
-  margin-left: 15px;
-  margin-right: auto;
-  width: 55px;
-  height: 35px;
-  cursor: pointer;
-}
-
-.MainNav ul {
-  list-style: none;
-  display: flex;
-}
-
-.NavContainer {
-  background-image: url("https://img.pilly.kr/main/v201911/cover-pc@2x.jpg");
-  position: relative;
-  overflow: hidden;
-  height: 100%;
-  min-height: 840px;
-  background-position-x: 50%;
-  background-position-y: 50%;
-  background-size: cover;
-}
