+++ src/components/Main-Chris/MainNav/MainNav.css
@@ -1,0 +1,32 @@
+.MainNav nav {
+  display: flex;
+  justify-content: space-between;
+  font-size: 15px;
+  font-weight: bold;
+  align-items: center;
+}
+
+.MainNav a {
+  margin: 20px;
+  align-items: center;
+  color: black;
+  text-decoration: none;
+  justify-content: end;
+}
+
+.MainNav img {
+  display: inline-block;
+  align-items: center;
+  margin-top: 15px;
+  margin-bottom: 15px;
+  margin-left: 15px;
+  margin-right: 15px;
+  width: 55px;
+  height: 35px;
+  cursor: pointer;
+}
+
+.MainNav ul {
+  list-style: none;
+  display: flex;
+}
