--- src/components/Main-Chris/MainScripts.js
@@ -1,5 +1,0 @@
-const MainScripts = () => {
-  return;
-};
-
-export default MainScripts;
