+++ src/components/Main-Chris/MainScripts/MainScripts.js
@@ -1,0 +1,18 @@
+const MainScripts = () => {
+  return (
+    <div className="MainScripts">
+      <h1>
+        내 몸에 필요한
+        <br />
+        영양성분 궁금하세요?
+      </h1>
+      <h3>
+        나만을 위한 맞춤영양제를 찾아보세요!
+        <br />
+        이미 1명이 추천 받았습니다
+      </h3>
+    </div>
+  );
+};
+
+export default MainScripts;
