--- src/pages/HomePage.js
@@ -1,13 +1,0 @@
-import MainNav from "../components/Main-Chris/MainNav";
-
-const HomePage = () => {
-  return (
-    <div className="NavContainer">
-      <div className="HomePage">
-        <MainNav />
-      </div>
-    </div>
-  );
-};
-
-export default HomePage;
