+++ src/pages/Test1/Test1.css
@@ -1,0 +1,10 @@
+.Test1 {
+  background-color: aqua;
+  position: relative;
+  overflow: hidden;
+  height: 100%;
+  min-height: 840px;
+  background-position-x: 50%;
+  background-position-y: 50%;
+  background-size: cover;
+}
