+++ src/pages/Test1/Test1.js
@@ -1,6 +1,9 @@
+import MainNav from "../../components/Main-Chris/MainNav/MainNav";
+
 const Test1 = () => {
   return (
-    <div>
+    <div className="Test1">
+      <MainNav />
       <h1>이곳은 테스트1 입니다.</h1>
     </div>
   );
