+++ src/pages/Test2/Test2.js
@@ -1,6 +1,9 @@
+import MainNav from "../../components/Main-Chris/MainNav/MainNav";
+
 const Test2 = () => {
   return (
     <div>
+      <MainNav />
       <h1>이곳은 테스트2 입니다.</h1>
     </div>
   );
